refresh, stable version

+++ pre_driver/roms_hcast_driver.sh
@@ -3,7 +3,8 @@ RA_ROOT=$2
 STRT_DATE_FULL=$3
 DT=$4
 CASE_NAME=$5
-
+INIT_RUN_FLAG=$6
+OFFSET_DAY=$7
 
 FCST_DAYS=1
 AEGIR_PROJ_PATH=${AEGIR_ROOT}/Projects/Aegir/
@@ -34,9 +35,11 @@ sed -i "s@CLMNAME_placeholder@CLMNAME == ${CLMFILE}@" $ROMS_IN
 sed -i "s@BRYNAME_placeholder@BRYNAME == ${BDYFILE}@" $ROMS_IN
 
 # relink ROMS icbc
-ICBC_LK=${AEGIR_PROJ_PATH}/ow_icbc/d01
-rm -f $ICBC_LK
-ln -sf ${ICBC_ROOT} ${ICBC_LK}
+if [ $INIT_RUN_FLAG == 1 ]; then
+    ICBC_LK=${AEGIR_PROJ_PATH}/ow_icbc/d01
+    rm -f $ICBC_LK
+    ln -sf ${ICBC_ROOT} ${ICBC_LK}
+fi
 
 # bug fix for bdy and clm files
-python3 ./roms_drv/roms_bdy_clm_time_bug_patch.py $AEGIR_PROJ_PATH $STRT_DATE_PACK 
+python3 ./roms_drv/roms_bdy_clm_time_bug_patch.py $AEGIR_PROJ_PATH $STRT_DATE_PACK $INIT_RUN_FLAG $OFFSET_DAY
